Address unspecified behavior in the WEBGL_multiview spec (#2176)
The native extension spec recommends an error in case the number of
views doesn't match between the framebuffer and the shader. This
should be required behavior in the WebGL version of the extension.

When the number of views specified in the active program is one,
it needs to be specified what happens with gl_ViewID_OVR. The simplest
option is to specify that it always evaluates to zero in this case.
Another alternative would be to forbid statically using gl_ViewID_OVR
in this case, but this could only be validated at link time.