Test uploading HTML data (canvas, image, etc.) to luminance and/or alpha textures. (#2451)
* Test uploading HTML data (canvas, image, etc.) to luminance and/or alpha textures.

The most extensive new tests are under conformance/textures/canvas and conformance2/textures/canvas, where semi-transparent cases have been added for all of the texture formats.

The canvas tests used to repeat multiple times, to act as regression tests for bugs in early WebGL implementations. The repetition has been kept for WebGL 1.0's RGB and RGBA internal formats, but removed for all other texture formats, to avoid dramatically increasing the tests' run time. Overall, this pull request will decrease the run time of the test suite.

Chrome, Firefox and Safari pass all of the new WebGL 1.0 tests with WebGL 1.0 contexts. Edge passes them, but generates errors about the UNPACK_COLORSPACE_CONVERSION_WEBGL parameter, which it does not support.

Firefox fails the new WebGL 1.0 luminance, alpha, and luminance_alpha tests when run against a WebGL 2.0 context, as tested on Linux and macOS. Emulation of these texture formats on the Core Profile is needed, as was done for Chrome. There is a preexisting failure of the RGB10_A2 tests in Firefox.

Chrome passes all of the tests with a forthcoming bug fix, and fails several of the tests without it.