Allow array length() method to be called on arbitrary expressions (#2493)
Reverse the expectations on a test that was based on an incorrect
interpretation of the spec. While the GLSL ES 3.00 spec says that
length is only allowed on "array names", it was always the intent to
allow it on arbitrary expressions that return an array. This has been
clarified in GLSL ES 3.20.

Since the revised test adds new requirements, it makes sense to
remove the test entirely from the 2.0.0 conformance suite.