Reject WEBGL_shared_resources.
At this point in the WebGL API's development it is not profitable to
invest the time to implement this extension. There are other higher
priorities, and all browsers are committed to making it possible to
draw WebGL-rendered canvases into other canvases quickly, which
essentially solves this problem.

Provides a bit more description than #826.

Removes test case added in #300.