Add a test for draw framebuffer completeness (#2671)
The test checks that the draw framebuffer can still be rendered to
even if the read framebuffer is incomplete.

The new test exposes a bug in Chrome 67. It seems to be specific to
ANGLE's D3D11 backend. Bug link: http://anglebug.com/2737

The latest stable version of Firefox is not affected by the bug.