integer-cubemap-specification-order-bug.html: add more test cases (#2806)
Makes this test more likely to catch the intended driver bug in the presence of attempted workarounds. (This caught that a few of my attempts at workarounds were not good enough. The eventual fix is (probably) to just preallocate POSITIVE_X, level 0 before everything else.)

These tests were generated by shuffling the initialization order
thousands of times and picking specifically the cases that failed
on the affected driver.

crbug.com/905003