Fix context-attribute-preserve-drawing-buffer.html (#2854)
There's no way to guarantee the compositor finished preparing the frame.

So we are gonna change the test to the following behavior:

Do an infinite for loop for `await nextFrame()` and keep doing readpixels and checking within the loop until we reach some time limit (say 0.5s). For testing `preserveDrawingBuffer: false` we could early terminate if the pixel data matches. And when testing `preserveDrawingBuffer: true` we could assert failure once the pixel doesn't match the expected ones.