Add RGTC drawing tests to existing S3TC test. (#3160)
The existing S3TC tests are very thorough. S3TC corresponds to the
BC1, BC2, and BC3 formats in DirectX. The RGTC extension adds BC4
and BC5, which are 1 and 2 channel variants of BC3. Rather than
build a new test, I just added BC4 and BC5 support to the existing
S3TC test.

RGTC was not fully tested before, and unsurprisingly there are bugs
in both Chrome and Firefox:
https://bugs.chromium.org/p/chromium/issues/detail?id=1136231
https://bugzilla.mozilla.org/show_bug.cgi?id=1669848