Add epsilon to float comparison (#3635)
Adds an epsilon when comparing float values in
conformance/extensions/webgl-depth-texture.html. As-is, it is possible
for tests to fail with output such as:

At 2,0, expected within [0.2,0.6], was 0.600
FAIL At 2,0, expected within [0.2,0.6], was 0.600

Removes the epsilon value added to certain parts of the expected min/max
values in conformance/extensions/webgl-depth-texture.html, instead using
only the epsilon applied when actually comparing values.